// src/features/pokemons/domain/enums/generations.enum.ts
export enum PokemonGenerations {
	Generation1 = 'generation-i',
	Generation2 = 'generation-ii',
	Generation3 = 'generation-iii',
	Generation4 = 'generation-iv',
	Generation5 = 'generation-v',
	Generation6 = 'generation-vi',
	Generation7 = 'generation-vii',
	Generation8 = 'generation-viii',
	Generation9 = 'generation-ix',
}